Inventory-backend/inventory/requirements.txt
# --- FastAPI stack ---
fastapi==0.128.0
starlette==0.50.0
uvicorn==0.40.0
anyio==4.12.1
h11==0.16.0
watchfiles==1.1.1
websockets==16.0
httptools==0.7.1
uvloop==0.22.1
click==8.3.1
idna==3.11

# --- Pydantic v2 ---
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings>=2.0
typing-extensions==4.15.0
typing-inspection==0.4.2
annotated-types==0.7.0

# --- Auth / Security ---
passlib==1.7.4
bcrypt==5.0.0
python-jose==3.5.0
rsa==4.9.1
pyasn1==0.6.2
ecdsa==0.19.1
six==1.17.0

# --- Database ---
sqlalchemy>=2.0
psycopg2-binary
alembic

# --- Config / Utils ---
python-dotenv==1.2.1
PyYAML==6.0.3

# --- Optional / Docs ---
annotated-doc==0.0.4
exceptiongroup==1.3.1
